requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.1.0
